Move kafka registration to a dedicated function
